bob: gru: kevin: Don't disable keyscanning in S0.
When the system is in S0, the EC should not disable the input devices.
The decision to ignore input devices is left to Chrome.  This commit
fixes lid_angle_peripheral_enable to only disable keyscanning when the
device is NOT in S0.  Additionally, the EC should not just cut off power
to the trackpad either.  This was fixed earlier for kevin and gru, but
not bob.

BUG=chrome-os-partner:61871
BRANCH=gru
TEST=Flash kevin; verify that key scanning is never disabled in S0.

Change-Id: I07008f022ccbc9588394db0541618065c45937a3
Signed-off-by: Aseda Aboagye <aaboagye@google.com>
Reviewed-on: https://chromium-review.googlesource.com/429190
Tested-by: Aseda Aboagye <aaboagye@chromium.org>
Reviewed-by: Gwendal Grignou <gwendal@chromium.org>
Commit-Queue: Aseda Aboagye <aaboagye@chromium.org>