g: harden upgrade validity checks for prod images
Let's not allow downloading older images when in prod mode.

When the received chunk is destined into RO or RW header, verify that
the chunk's version is not lower than the current running version.

Also, if the chunk is not properly aligned with the header, verify
that it does not overlap with the header in any way.

BRANCH=none
BUG=chrome-os-partner:63098
TEST=verified that older images are rejected by prod images, and newer
     and current level are accepted.

     Verified that dev images still allow to downgrade.

Change-Id: I19c74f1d1bb5469cc935293a5841405149a968f6
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/448702
Reviewed-by: Randall Spangler <rspangler@chromium.org>