servo_v4: pd: Updated CC_NC and CC_RA macros
Previous boards that implemented tcpc layer on chip didn't support
variable Rp values. However, servo_v4 can present any of the 3
possible Rp values and therefore the voltage thresholds that are used
to determine a no-connect or Ra attach status need a way to be set
based on the Rp value that's current attached to a given CC line.

- Added port and cc line selection to both the CC_NC and CC_RA macros
  and now check if they are already defined before being defined in
  usb_pd_tcpc.c.
- Defined each of these macros in board.h to use a function that's
  able to select the threshold based on the current Rp configuration.

BUG=b:35586526
BRANCH=servo
TEST=Tested with servo_v4 against Electro and verified that it
connects when a charger is and is not connected to CHG port which
exercises the differnt Rp combinations that servo_v4 presents.

Signed-off-by: Scott <scollyer@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/451962
Commit-Ready: Scott Collyer <scollyer@chromium.org>
Tested-by: Scott Collyer <scollyer@chromium.org>
Reviewed-by: Vincent Palatin <vpalatin@chromium.org>
Change-Id: I1a31e430c0f290486f0fa8a50bdafdddf20d23ca
Reviewed-on: https://chromium-review.googlesource.com/506957
Reviewed-by: Nick Sanders <nsanders@chromium.org>