BD9995X: Disable input current limiting for VBAT < VSYSREG_SET
The initial value for input current limit is set to
CONFIG_CHARGER_INPUT_CURRENT which is typically 512 mA. In deeply
discharged battery cases (Vbat < 5.8V), the 512 mA input current limit
can cause VSYS to collapse which in turn causes the EC to
reset. Depending on how discharged the battery is, the EC may remain
off until the external charger is disconnected and reconnected again,
or it may undergo a number of reset cycles, each time charging the
battery just a little, until Vbat becomes > ~5.8V and the charger is
able to stabilize. When the charger type is determined, either from
BCD detection, or Type C/USB PD, the input current limit is set to the
appropriate level.

In order to avoid the issue described above, this CL sets a bit in the
VIN_CTRL_SET register which will disable the input current limit in cases
Where the VBAT is less than the VSYSREG_SET value.

BUG=b:35648317
BRANCH=none
TEST=Manually tested on Electro.
     a. With Zinger attached DUT boots without the battery after
        plugging in AC
     b. DUT boots from cut-off battery
     c. With Zinger attached DUT boots from cold-reset without the
        battery
     d. With no battery & DCP charger, anti-collapse occurs,
        input current is limited to 512mA & the DUT is
        power-up inhibited.

Tested also on Eve with signal wires attached to both PPVAR_VSYS,
PP3300_DSW, and Vbat. Verified that on certain boards (some board to
board variation) that PPVAT_VSYS would collapse when the input current
limit was set to CONFIG_CHARGER_INPUT_CURRENT. Then after adding this
CL, verifed on the scope that the collapse of PPVAR_VSYS no longer
occurred.

Change-Id: I63d7b236fe9b513b353b6bcb6ad19ac498900de9
Signed-off-by: Duncan Laurie <dlaurie@google.com>
Original-Commit-Id: ce65199e5ed75bf8ab4cc3a4f9abae85f63f1a02
Original-Change-Id: Ief9960550f988e69ab4637db85450e91c70d3b51
Original-Signed-off-by: Vijay Hiremath <vijay.p.hiremath@intel.com>
Original-Reviewed-on: https://chromium-review.googlesource.com/456049
Original-Commit-Ready: Vijay Hiremath <vijay.p.hiremath@intel.corp-partner.google.com>
Original-Tested-by: Scott Collyer <scollyer@chromium.org>
Original-Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Original-Reviewed-by: Vijay Hiremath <vijay.p.hiremath@intel.corp-partner.google.com>
Reviewed-on: https://chromium-review.googlesource.com/524410