usb_updater: fix -b option
We want the -b option to be parsed the same way as the -u option,
because usb_updater should handle reading the file the same way. This
changes binver hasarg to be 0, so that it doesn't eat the filename.

BUG=none
BRANCH=cr50
TEST=usb_updater --binver $IMAGE

Change-Id: I0b868bc5d316e5fb42fc34bc746bbee868d20630
Signed-off-by: Mary Ruthven <mruthven@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/465490
Reviewed-by: Aaron Durbin <adurbin@chromium.org>
(cherry picked from commit 24fd8060fd27f0306d5d3f91faa09d28ed892772)
Reviewed-on: https://chromium-review.googlesource.com/470486