skylake: Add workaround for boards that cannot save reset flags
Some hardware has an issue where the reset flags are lost on power cycle
because the EC backup ram loses power.  This causes the flag to not power
on the AP (ap-off) to be lost.

In order to pass FAFT it is required that boards support this flag, so
this commit adds a workaround where the skylake chipset code will call into
the board to ask if it has working reset flags and if not it will skip the
PMIC reset if the "ap-off" flag has been set.

The "ap-off" flag is purely for testing, it is not possible for users to
do this without having access to the EC console.  (which is currently not
possible at all with CCD unless you can also build a debug cr50 image)

BUG=b:38187362,b:35585876
BRANCH=none
TEST=manual testing on Eve: execute 'reboot ap-off' and ensure that the
AP does not power on.  Also ensure that 'dut-control power_state:rec' works
as expected and does not power off at the recovery screen due to a power
button press.

Change-Id: Ic071e8c68ed38a3820ca49e3d23503f10e0f68ef
Signed-off-by: Duncan Laurie <dlaurie@google.com>
Original-Commit-Id: dcedeab2cceac2eb2e5f1e29639cf0271283a18f
Original-Change-Id: If11e17179e9173509b9a6ae1ef0d94a50ba181d0
Original-Signed-off-by: Duncan Laurie <dlaurie@google.com>
Original-Reviewed-on: https://chromium-review.googlesource.com/514503
Original-Reviewed-by: Furquan Shaikh <furquan@chromium.org>
Original-Reviewed-by: Nicolas Boichat <drinkcat@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/524372