cr50: change the power button handling
Enable permanently the power button interrupt and discard the unneeded
presses, rather than enabling the interrupt on demand.
This is a preparatory work for the U2F feature which is using the power
button as the user physical presence sensing. It will need to be able to
detect all the button touches.

Signed-off-by: Vincent Palatin <vpalatin@chromium.org>

BRANCH=cr50
BUG=b:35545754
TEST=with follow-up CLs, run U2FTest on Eve.

Change-Id: I36536a48ca7beebaf3c6d12d1ad3e991897ba82f
Signed-off-by: Duncan Laurie <dlaurie@google.com>
Original-Commit-Id: ea274f2722f432f8f9f4b3f83336dd25d0f6d4f7
Original-Change-Id: Ifd90970932684d31ad7687db260bafa65a56e854
Original-Reviewed-on: https://chromium-review.googlesource.com/518138
Original-Commit-Ready: Vincent Palatin <vpalatin@chromium.org>
Original-Tested-by: Vincent Palatin <vpalatin@chromium.org>
Original-Reviewed-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/524397