cr50: Have CCD_MODE_L respect "ccd keepalive".
If CCD was set to the "keepalive" mode, removing the debug accessory
would cause the CCD_MODE_L pin to be pulled high, ignoring the keepalive
state.  CCD was still kept alive, but the state reflected on the pin
didn't indicate so.

This commit changes the behaviour such that the CCD_MODE_L pin is set up
as an input only when CCD keepalive is not enabled.

BUG=None
BRANCH=cr50
TEST=Plug in debug accessory and remove debug accessory.  Verify that
CCD_MODE_L is high after removal.
TEST=Enable ccd keepalive.  Remove debug accessory.  Verify that
CCD_MODE_L is still low.

Change-Id: Ifaae6889f011baa87f487010c48d261545aa02a6
Signed-off-by: Duncan Laurie <dlaurie@google.com>
Original-Commit-Id: 01b811f56d384062256bb7d940e7b84c02b984f9
Original-Change-Id: I407994d5c394a717d6e1a87f283f6441bd26bf55
Original-Signed-off-by: Aseda Aboagye <aaboagye@google.com>
Original-Reviewed-on: https://chromium-review.googlesource.com/520603
Original-Commit-Ready: Aseda Aboagye <aaboagye@chromium.org>
Original-Tested-by: Aseda Aboagye <aaboagye@chromium.org>
Original-Reviewed-by: Mary Ruthven <mruthven@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/524392