npcx: i2c: change i2c settings for better timing.
The setup time of original i2c setting which i2c freq is 1MHz is closed
to the margin of the spec. This CL improves the setup time with a new
setting. (For example, in npcx evb, measured setup time now is about
480 ns.)
This CL also removes the timing settings of higher i2c source clock
frequencies which are not used so far to save the code size.

BRANCH=none
BUG=b:38217035
TEST=No build error for make buildall; run stress test with i2cxfer and
i2cscan on gru with 400K and 1MHz i2c freq.

Change-Id: I8d692fbe436e182421115a741c7ce9796d8fbaea
Signed-off-by: Duncan Laurie <dlaurie@google.com>
Original-Commit-Id: ee089e62a7a53cad95b99c66f7a6f8fd4e2ab1c5
Original-Change-Id: I5428a7dab1d935fd428ee9012604813e752cead8
Original-Signed-off-by: CHLin <CHLIN56@nuvoton.com>
Original-Reviewed-on: https://chromium-review.googlesource.com/527739
Original-Commit-Ready: CH Lin <chlin56@nuvoton.com>
Original-Tested-by: CH Lin <chlin56@nuvoton.com>
Original-Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/542322