KBL/SKL: Add wait between DSW_PWROK and PWRBTN
This patch adds wait between DSW_PWROK and PWRBTN assert. It is
required to be 95 msec or longer for Kaby Lake and Sky Lake.
Refer to the timing diagram for G3 to S0 on Sky Lake or Kaby Lake
platform design guide for details.

BUG=b:62584658
BRANCH=none
TEST=On Fizz, measured time between DSW_PWROK high and PWRBTN assert
for 1:AC plug-in, 2:recovery+power press, 3: reboot ec command.

Change-Id: If067e5f1ad18647f6211c45fbb09cd153f7cf939
Signed-off-by: Duncan Laurie <dlaurie@google.com>
Original-Commit-Id: 627f3a8ac28751f549fbdb5266f8ccee417d339d
Original-Change-Id: I89a14ac9a49e20a332bd662d90be62f8ea23b003
Original-Signed-off-by: Daisuke Nojiri <dnojiri@chromium.org>
Original-Reviewed-on: https://chromium-review.googlesource.com/534901
Reviewed-on: https://chromium-review.googlesource.com/542324