bob,gru,scarlet: Assert SYS_RST_L by default.
In order for cr50 to be able to detect when the AP is in its S5 power
state, the power sequencing driver was changed to ensure that SYS_RST_L
remains asserted in S5 and then is deasserted on the transition up,
instead of just pulsing during the S5->S0 transition.

To make sure that the signal remains asserted in S5, the GPIOs are
Changed to be initalized to low coming out of EC reset.

Since the power sequencing driver changes affect all RK3399 boards (with
the exception of kevin), all the GPIOs are adjusted in order to be
consistent.

BUG=b:35647982
BRANCH=gru
TEST=make -j buildall
TEST=Flash a bob, verify it's able to boot up and shutdown okay.  Warm
resets are also functional.

Change-Id: I0ce8b365e12f016c731e6b1a0ee29e43e31f5c59
Signed-off-by: Aseda Aboagye <aaboagye@google.com>
Reviewed-on: https://chromium-review.googlesource.com/538984
Reviewed-by: Vadim Bendebury <vbendeb@chromium.org>
Tested-by: Vadim Bendebury <vbendeb@chromium.org>
Commit-Queue: Aseda Aboagye <aaboagye@chromium.org>