npcx: Clear IBBR to have BBRAM back to be functional
In the CL:505861, we print the warnning message to indicate that the VBAT
has ever dropped if the IBBR bit is set but do not clear the IBBR. This
forbid the access to BBRAM until EC power-on reset.
In this CL, we clear the IBBR bit to make BBRAM work again without the
need of EC power-on reset.

BRANCH=none
BUG=none
TEST=No build error for make buildall; Check warining messages are
printed and IBBR bit is cleared.

Change-Id: Id19022dce6bc3fb98b2ca5b846b065ae4883e1f2
Signed-off-by: Duncan Laurie <dlaurie@google.com>
Original-Commit-Id: 5e5788f3cac7fecd45072807bb6a79ce2b767961
Original-Change-Id: I58a0370c1c496b3c1208d9d5ac6b55c4d66fe8b6
Original-Signed-off-by: CHLin <CHLIN56@nuvoton.com>
Original-Reviewed-on: https://chromium-review.googlesource.com/542976
Original-Commit-Ready: CH Lin <chlin56@nuvoton.com>
Original-Tested-by: CH Lin <chlin56@nuvoton.com>
Original-Reviewed-by: Shawn N <shawnn@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/565816