eve: Minor changes to LED patterns
The LED pattern for lowest battery level is supposed to indicate that
there isn't enough battery present to boot the AP. The pattern_tbl
array had this value set to 2. However, it's better to tie this to the
constant defined in board.h so that it's consistent with when the AP
will be allowed to boot. In addition, adjusted the battery level
ranges to work correctly with the <= check.

BUG=b:35584895
BRANCH=eve
TEST=Used 'battfake' console command to verify that the patterns
changed at the expected battery charge levels.

Change-Id: Ifbf68229d8b373f1da60b52355904f0476def0cf
Signed-off-by: Duncan Laurie <dlaurie@google.com>
Original-Commit-Id: dccb0ed867c7409e06faf7b9b353ddb235c43ee0
Original-Change-Id: Ie85a448ad4803be7e612c382b8e60fa42356d257
Original-Signed-off-by: Scott Collyer <scollyer@google.com>
Original-Reviewed-on: https://chromium-review.googlesource.com/562542
Original-Commit-Ready: Scott Collyer <scollyer@chromium.org>
Original-Tested-by: Scott Collyer <scollyer@chromium.org>
Original-Reviewed-by: Todd Broch <tbroch@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/567646