eve: Delay 1s before powering AP after battery disconnect
When coming out of battery disconnect the battery may take some time
to provide power, and we should wait for that before we try to boot
the AP or it may brown-out unexpectedly.

To be safe this adds a 1 second delay, which only ever happens when
first booting after recovering a disconnected battery.

BUG=b:63360549
BRANCH=eve
TEST=manual testing with cut-off battery, see that it wakes up and
then after 1s allows the AP to boot:
[0.006188 battery not found]
[1.486592 battery woke up]
[2.487140 battery will now allow booting]

Change-Id: I030f57ec3a2f8b26d72fc9fff70fe449fe00b1c5
Signed-off-by: Duncan Laurie <dlaurie@google.com>
Reviewed-on: https://chromium-review.googlesource.com/584208
Reviewed-by: Shawn N <shawnn@chromium.org>
Reviewed-by: Scott Collyer <scollyer@chromium.org>
Reviewed-by: Furquan Shaikh <furquan@chromium.org>