eve: Modify LED behavior to show double tap patterns when charging
Previous LED behavior was to have only the LED on side that was
charging show white or green. The other side LED was always off.

This CL modifies the behavior so that when double tap patterns are
allowed and a charger is connected to display the double tap pattern
on the opposite side LED which was previously off.

In addition, when the device is charging, if the power level is such
that AP can't boot, the opposite side LED will display the BLINK_RED
pattern continuously as it would duirng a double tap event.

In order to support the ability to have different LED patterns
simultaneously, the LED code was refactored so that static variables
used to manage the patterns and varialbes required to manage color
gradient transitions were grouped together in to a static led
descriptor. In addition, the led_manage_pattern and change_color
functions had to be modified to deal with each LED independently.

BUG=b:35584895
BRANCH=eve
TEST=Manual
Used EC console command 'battfake' to force different charge levels
and verified the correct patterns for double tap events. Confirmed
that double tap patterns are displayed on the non-charging port when a
charger is connected. Tested connecting and removing charger while
observing non-charging port.

Change-Id: I05e1671deada761388bf898096ed7d3ae6f8da0f
Signed-off-by: Scott Collyer <scollyer@google.com>
Reviewed-on: https://chromium-review.googlesource.com/611781
Tested-by: Scott Collyer <scollyer@chromium.org>
Reviewed-by: Duncan Laurie <dlaurie@google.com>
Commit-Queue: Scott Collyer <scollyer@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/2320255
Tested-by: Patryk Duda <pdk@semihalf.com>
Commit-Queue: Patryk Duda <pdk@semihalf.com>
Reviewed-by: Jett Rink <jettrink@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/2396688
Reviewed-by: Edward Hill <ecgh@chromium.org>
Commit-Queue: Edward Hill <ecgh@chromium.org>
Tested-by: Edward Hill <ecgh@chromium.org>