usb_pd_protocol: Req SNK Cap if not received yet.
In the SRC_READY state, we'll only request sink caps if we haven't
received them yet and only if its the first transition to the state.
However, we also don't send any PD traffic in that state if there's an
incoming message in order to prevent a collision.  This could create a
scenario where upon entry to the SRC_READY state, a message is incoming.
When this occurs, we never request the sink caps.

This commit simply removes the condition that we may only request sink
caps on the first transition to the SRC_READY state.

BUG=b:64037926
BRANCH=gru
TEST=Flash kevin; Connect to a DR port partner; Verify that sink caps
are requested even after the first transition to the SRC_READY state.

Change-Id: Ie5f62df548b6e90930c6b3cbb18a786fb063c10e
Signed-off-by: Aseda Aboagye <aaboagye@google.com>
Original-Change-Id: I6bc9ad01d45e6584a7a14b28806ae4872a22d98f
Original-signed-off-by: Aseda Aboagye <aaboagye@google.com>
Original-Reviewed-on: https://chromium-review.googlesource.com/611320
Original-Commit-Ready: Aseda Aboagye <aaboagye@chromium.org>
Original-tested-by: Aseda Aboagye <aaboagye@chromium.org>
Original-reviewed-by: Shawn N <shawnn@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/614652
Reviewed-by: Aseda Aboagye <aaboagye@chromium.org>
Commit-Queue: Aseda Aboagye <aaboagye@chromium.org>
Tested-by: Aseda Aboagye <aaboagye@chromium.org>