anx3429: force chip reset on PD_RESUME
we need to properly restart the anx3429 after a firmware update.
simply initializing the chip doesn't seem to get it to reload its
firmware - at least not the portion of the chip that implements the
firmware version register.  so, we explicitly power down and reset the
chip before reinitializing it to force it to run the new firmware.

the chip also needs a 10ms "off" time so the reset is properly seen by
the chip, so i did a light refactoring of the code paths that reset
the anx3429.

TEST=used 2 different firmware blobs and verified it switches between
	them during software sync.

BRANCH=none
BUG=b:35586895

(cherry picked from commit c1e5671e561c82c9532cb29b17fc7cf4061ab20a)
Reviewed-on: https://chromium-review.googlesource.com/631976

Change-Id: I321b5bc365b477c9b8754417b01fd03921fc648b
Signed-off-by: Caveh Jalali <caveh@google.com>
Reviewed-on: https://chromium-review.googlesource.com/644771
Reviewed-by: Shawn N <shawnn@chromium.org>