anx3429: force chip reset on PD_RESUME
we need to properly restart the anx3429 after a firmware update.
simply initializing the chip doesn't seem to get it to reload its
firmware - at least not the portion of the chip that implements the
firmware version register.  so, we explicitly power down and reset the
chip before reinitializing it to force it to run the new firmware.

the chip also needs a 10ms "off" time so the reset is properly seen by
the chip, so i did a light refactoring of the code paths that reset
the anx3429.

TEST=used 2 different firmware blobs and verified it switches between
        them during software sync.

BRANCH=none
BUG=b:35586895

Change-Id: I11447c4039108a7530028c57cd5ce5ef6a78912a
Signed-off-by: Duncan Laurie <dlaurie@google.com>
Original-Commit-Id: c1e5671e561c82c9532cb29b17fc7cf4061ab20a
Original-Change-Id: I967898dd906f21bdc5bc4ce9c1dff9f873d198c1
Original-Signed-off-by: Caveh Jalali <caveh@google.com>
Original-Reviewed-on: https://chromium-review.googlesource.com/631976
Reviewed-on: https://chromium-review.googlesource.com/914659