anx3429: read chip firmware version during init
fetching the chip firmware version toward the end of the chip
anx74xx_tcpm_init() sequence is a good place to do this.  we need this
info in any case and this is a safe place to access device registers
and cache the values.  subsequent chip firmware queries typically
return the cached value.  also, tcpci_tcpm_init() is already
structured this way.

TEST=verified with follow-up CL that firmware update succeeds and new
        version is reported

BRANCH=none
BUG=b:35586895

Change-Id: If3a7ac37dc978db655b5dfb9c9df0756d08162d9
Signed-off-by: Duncan Laurie <dlaurie@google.com>
Original-Commit-Id: e4997a631e129df3d171da16de9504810891744c
Original-Change-Id: Ic3fd07bbf8a220bfd506d59d8a1f3ea25b14e94c
Original-Signed-off-by: Caveh Jalali <caveh@google.com>
Original-Reviewed-on: https://chromium-review.googlesource.com/634513
Original-Reviewed-by: Shawn N <shawnn@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/914657