pd: Allow deep sleep in SRC_DISCOVERY
If the PD state machine remains in SRC_DISCOVERY for an extended period
of time, it's likely that a non-PD USB peripheral is attached. In this
case, we don't need to inhibit deep sleep, since we're not likely to
receive PD packets.

This change will cause us to enter deep sleep slightly more
aggressively, not inhibiting deep sleep until source caps are received
or replied with GoodCRC by the port partner. We can accommodate
additional task latency up to this point, since the spec calls for
source caps to be sent up to 50 times before failure.

BUG=b:35582718,chromium:763002
TEST=Test with `sleepmask 1` on kevin.
- Go to S3 with USB-C flash drive plugged, verify `sleepmask` shows 0.
- Go to S3 with zinger + USB C flash drive plugged
   - Unplug zinger, verify `sleepmask` shows 0.
   - Plug zinger, verify PD negotiates to 20V @ 2A.
   - Plug OEM kevin charger, verify same.
BRANCH=gru

Change-Id: Iaedb8ed124a8aa614fabb58b7aed151ae7003dad
Signed-off-by: Duncan Laurie <dlaurie@google.com>
Original-Commit-Id: 9abb9f762efb3f386befc74e90275fe8789e3bd2
Original-Signed-off-by: Shawn Nematbakhsh <shawnn@chromium.org>
Original-Change-Id: Ib8e1bc94bdbcfddea004d572edf1ccadc8c8c1ce
Original-Reviewed-on: https://chromium-review.googlesource.com/655919
Original-Commit-Ready: Shawn N <shawnn@chromium.org>
Original-Tested-by: Shawn N <shawnn@chromium.org>
Original-Reviewed-by: Daisuke Nojiri <dnojiri@chromium.org>
Original-Reviewed-by: Vincent Palatin <vpalatin@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/690177