[EXPERIMENTAL] Add packetized console mode
Add hooks to provide console commands and read their answers in a
'packetized' fashion, ie it's the command is in a single delimited in a
single packet and doesn't need to be extracted from a flow of console
characters.

Signed-off-by: Vincent Palatin <vpalatin@chromium.org>

BRANCH=twinkie
BUG=none
TEST=with the following CLs, use the console with the packetized USB
endpoint on Twinkie.

Change-Id: Ie02db092af6f6f37c1d575bf5746183a64baafd1
Reviewed-on: https://chromium-review.googlesource.com/702423
Tested-by: Vincent Palatin <vpalatin@chromium.org>
Trybot-Ready: Vincent Palatin <vpalatin@chromium.org>
Commit-Queue: Vincent Palatin <vpalatin@chromium.org>
Reviewed-by: Vincent Palatin <vpalatin@chromium.org>