twinkie: add USB interface for packetized console commands
Add a new USB interface for commands with a pair of bulk endpoints.
It can be used to send console commands but contrary to the regular
'console' USB interface, the command/response are properly split by USB
packets (not free flowing as a regular console) and the response only
includes the traces emitted during the command execution on the
CC_COMMAND channel.
This USB interface is not declared as Google serial class contrary the
other regular USB console interface, so it can be used easily by various
tools (e.g. WebUSB or Sigrok) without being pre-empted by the kernel
usb-serial-simple driver.

The protocol on this interface (number 2) is the following:
- write a full (and single) EC console command on the OUT endpoint 2
(without the carriage return, up to 64 characters)
- poll the IN endpoint 0x82, when the command is done executing, it will
  return one or several packets (depending on the size) with the
  response.
(keep reading packets as long as they are 64-byte long, the last one
will be a partial one, possible 0-byte long if the response size is a
multiple of 64 bytes).

Signed-off-by: Vincent Palatin <vpalatin@chromium.org>

BRANCH=twinkie
BUG=none
TEST=on Twinkie, run a tight loop of 'tw vbus' command to test
robustness.

Change-Id: Ibb4b25d7af053d8452924ce0d6bbcb0685ed7752
Reviewed-on: https://chromium-review.googlesource.com/702424
Tested-by: Vincent Palatin <vpalatin@chromium.org>
Trybot-Ready: Vincent Palatin <vpalatin@chromium.org>
Commit-Queue: Vincent Palatin <vpalatin@chromium.org>
Reviewed-by: Vincent Palatin <vpalatin@chromium.org>