pd: Add pd_capable() to check PD capability of partner port
It's undesirable to do BC1.2 detection on power swap, so add a function
to check if the partner port is known to be PD-capable.

BUG=chromium:780905
BRANCH=gru
TEST=With subsequent CL, attach USB-PD phone capable of role swap.
Verify USB 2.0 device is enumerated on plug, and not re-enumerated
through a series of "pd # swap power" commands on the EC console. Also
verify BC1.2 charging and PD charging are still functional on kevin.

Change-Id: I4b775506beadb66e170b34a8acedf69a8a5dcf3d
Signed-off-by: Duncan Laurie <dlaurie@google.com>
Original-Commit-Id: fee1bde58def62ff7cbd40e060844c0cc496a032
Original-Signed-off-by: Shawn Nematbakhsh <shawnn@chromium.org>
Original-Change-Id: Ifa75c94e9758d3e407492bbda6fc52ed7bc378fa
Original-Reviewed-on: https://chromium-review.googlesource.com/755877
Original-Commit-Ready: Shawn N <shawnn@chromium.org>
Original-Tested-by: Shawn N <shawnn@chromium.org>
Original-Reviewed-by: Vincent Palatin <vpalatin@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/914648