Snappy: Extend the watchdog timeout to 500ms after wakeup from hibernate
Currently, when EC enters hibernation because of critical battery level,
upon AC plug-in, EC will wake up from hibernation and execute the
watchdog reset immediately. However, on Snappy, we observed
that there is a drop on the PP1800_EC power rail after AC is applied
when VBAT is ~5.8V. If the drop happens when the EC booter accesses
the SPI flash, it will cause EC to hang because booter may not find a
valid image and go to an endless loop. This CL extends the watchdog
reset timeout to 500 ms. This will delay the timing that EC's booter
accesses the SPI flash to the period where PP1800_EC is stable.

BUG=b:70183568
BRANCH=reef
TEST=No build errors for "make buildall".
TEST=Measure the F_CS pin and PP1800_EC after plugging AC;
Make sure the EC start to access SPI flash after PP1800_EC drops.
Snappy EC can reboot succesfuly and show the firmware screen
10 out of 10 attempts.

Change-Id: I49fccdc3e588cdc24759b137af52af355d9bb47f
Signed-off-by: CHLin <CHLIN56@nuvoton.com>
Reviewed-on: https://chromium-review.googlesource.com/862545
Tested-by: Daisuke Nojiri <dnojiri@chromium.org>
Reviewed-by: Vincent Palatin <vpalatin@chromium.org>
Commit-Queue: Daisuke Nojiri <dnojiri@chromium.org>