meowth: Add support for base detection.
Meowth has two analog detection pins with which it monitors to determine
the base status: the attach, and detach pins.  When the voltages cross a
certain threshold, after some debouncing, the base is deemed connected.
Meowth then applies the base power and monitors for power faults from
the eFuse as well as base disconnection.  Similarly, once the voltages
cross a different threshold, after some debouncing, the base is deemed
disconnected.  At this point, Meowth disables the base power.

BUG=b:69140200
BRANCH=None
TEST=Verify that we can detect a base attach and detach.
TEST=Verify that when base is attached, tablet mode goes to 0, and goes
to 1 when the base is detached.

Change-Id: I2a069fc450c2232c7eae02b8e7459d38a049668a
Signed-off-by: Aseda Aboagye <aaboagye@google.com>
Reviewed-on: https://chromium-review.googlesource.com/952491
Reviewed-by: Aseda Aboagye <aaboagye@chromium.org>
Commit-Queue: Aseda Aboagye <aaboagye@chromium.org>
Tested-by: Aseda Aboagye <aaboagye@chromium.org>