bd9995x: Add chop mode control based on AP power state and CHG_EN
This CL adds chopper mode control which is gated on both the AP power
state and CHG_EN setting of the bd9995x. In chopper mode the bd9995x
switching converter consumes less energy, but can only be safely used
under light load conditions. If AP power != S0 and CHG_EN is 0, then
chopper mode will be enabled.

Because we disable the switching converter when the battery reaches
FULL, that effectively means that chopper mode will only be active
when the AP power state != S0 and when the battery is removed/cutoff.

This CL also removes a check in charger_set_current that would force
the current command to be >= 512 mA. This would prevent CHG_EN from
ever being 0 even when battery is removed. This check is not required
because the following section of code will always make sure that
charging is disabled if current command is 0 prior to setting the
command value to the ICHG register.

BUG=b:72979303
BRANCH=coral
TEST=On a coral device, forced AP power state to off with ec console
command 'aps' the issued 'cutoff'. Verifed that chopper mode was
enabled.

[409.054257 Battery cut off]
> [409.263562 charge problem: batt params, 0x0 -> 0x7fc after 409.263545s]
[409.265881 charge_request(0mV, 0mA)]
[409.380555 bd9995x: Chop Mode change, new val = 1]

Change-Id: I61bd52a0865502b2919a8a9b1274b1eb26c2e878
Signed-off-by: Scott Collyer <scollyer@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1010909
Tested-by: Scott Collyer <scollyer@chromium.org>
Reviewed-by: Todd Broch <tbroch@chromium.org>
Reviewed-by: Casper Chang <casper_chang@wistron.corp-partner.google.com>
Commit-Queue: Scott Collyer <scollyer@chromium.org>
Trybot-Ready: Scott Collyer <scollyer@chromium.org>